Move some code of Main class in new classes
