+++ src/main/java/com/alphatica/genotick/genotick/PopulationPrinter.java
@@ -1,0 +1,72 @@
+package com.alphatica.genotick.genotick;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Modifier;
+import java.util.ArrayList;
+import java.util.List;
+
+import com.alphatica.genotick.population.Population;
+import com.alphatica.genotick.population.PopulationDAOFileSystem;
+import com.alphatica.genotick.population.RobotInfo;
+
+class PopulationPrinter
+{
+    public static void printPopulation(String path) throws IllegalAccessException {
+        PopulationDAOFileSystem dao = new PopulationDAOFileSystem(path);
+        Population population = PopulationFactory.getDefaultPopulation(dao);
+        showHeader();
+        showRobots(population);
+    }
+
+    private static void showRobots(Population population) throws IllegalAccessException {
+        for(RobotInfo robotInfo: population.getRobotInfoList()) {
+            String info = getRobotInfoString(robotInfo);
+            System.out.println(info);
+        }
+    }
+
+    private static String getRobotInfoString(RobotInfo robotInfo) throws IllegalAccessException {
+        StringBuilder sb = new StringBuilder();
+        Field [] fields = robotInfo.getClass().getDeclaredFields();
+        for(Field field: fields) {
+            field.setAccessible(true);
+            if(!Modifier.isStatic(field.getModifiers())) {
+                Object object = field.get(robotInfo);
+                if(sb.length() > 0) {
+                    sb.append(",");
+                }
+                sb.append(object.toString());
+            }
+        }
+        return sb.toString();
+    }
+
+    private static void showHeader() {
+        Class<RobotInfo> infoClass = RobotInfo.class;
+        List<Field> fields = buildFields(infoClass);
+        String line = buildLine(fields);
+        System.out.println(line);
+    }
+
+    private static List<Field> buildFields(Class<?> infoClass) {
+        List<Field> fields = new ArrayList<>();
+        for(Field field: infoClass.getDeclaredFields()) {
+            if(!Modifier.isStatic(field.getModifiers())) {
+                fields.add(field);
+            }
+        }
+        return fields;
+    }
+
+    private static String buildLine(List<Field> fields) {
+        StringBuilder sb = new StringBuilder();
+        for (Field field : fields) {
+            if (sb.length() > 0) {
+                sb.append(",");
+            }
+            sb.append(field.getName());
+        }
+        return sb.toString();
+    }
+}
+
